Split prepare.py into two & rm hardcoded nodata vals
